+++ packages/compat/plugin-swc/package.json
@@ -37,6 +37,7 @@
     "@rsbuild/core": "workspace:*",
     "@rsbuild/webpack": "workspace:*",
     "@types/lodash": "^4.17.5",
+    "@types/semver": "^7.5.8",
     "typescript": "^5.4.2",
     "webpack": "^5.92.0"
   },
