+++ packages/core/src/plugins/basic.ts
@@ -1,5 +1,5 @@
 import path from 'node:path';
-import { isProd, isUsingHMR } from '@rsbuild/shared';
+import { isProd } from '@rsbuild/shared';
 import { TARGET_ID_MAP } from '../constants';
 import type { NormalizedConfig, RsbuildPlugin } from '../types';
 
@@ -53,7 +53,7 @@ export const pluginBasic = (): RsbuildPlugin => ({
         // set minimize to allow users to disable minimize
         chain.optimization.minimize(isMinimize);
 
-        const usingHMR = isUsingHMR(config, { target, isProd });
+        const usingHMR = !isProd && config.dev.hmr && target === 'web';
 
         if (usingHMR) {
           chain
