+++ packages/core/src/provider/createCompiler.ts
@@ -6,7 +6,6 @@ import {
   color,
   isDev,
   isProd,
-  prettyTime,
 } from '@rsbuild/shared';
 import { rspack } from '@rspack/core';
 import type { StatsCompilation } from '@rspack/core';
@@ -16,6 +15,7 @@ import {
   getStatsOptions,
   isSatisfyRspackVersion,
   onCompileDone,
+  prettyTime,
   rspackMinVersion,
 } from '../helpers';
 import { logger } from '../logger';
